Add `event-stage-form` to `setEventStage`
- Add component `event-stage-form` which displays a button for each of
the 3 possible event stages. When any of these buttons is clicked, the
button id, which matches the property name of a boolean field on the
event object, is sent as a parameter to `setEventStage`, which is
passed up through parent controllers until it reaches the route.
- At the `admin.js` route, the boolean values of each stage property are
updated so only one the selected one is now true, and the event record
is saved to the database.
- In `event-stage-form.hbs`, the template checks whether each boolean is
true and adds a `selected` class to the relevant button (all of which
also have a `stage` class regardless)
- In `app.scss`, items with both `.stage` and `.selected` classes get
highlighted in green so it's quickly visible which stage is selected