Add `timeslots` form / `createTimeslot` action
- Create `timeslots-form` in admin panel which asks how many rooms and
times are available. On submissions, the `submitTimeslots` action is
called in `timeslots-form.js`, where the system adjusts the
`timeslotsRequest` POJO object to have arrays of dummy room names, start
times, and end times of appropriate length, then loops through all the
combinations and pushes a relevant pojo onto the array `newTimeslots`
and triggers the `createTimeslot` action for each to be sent up to the
admin router where it attempts to create a new timeslot in the database.
- Currently `createTimeslot` isn't working because Ember tries to
send the `event` field as `null`