Add `new-timeslots display`, delete, fix create
- Add `new-timeslots-display` component, called from within
`timeslots-form` when timeslots exist to display them
- Add a `deleteAllTimeslots` function that fetches all timeslots, loops
through and destroys each of them so the admin user can start over with a
new number of rooms or session times.
- Update the `timeslot` model to look for `event_id` as an attribute
(now hardcoded to 1 in the `createTimeslots` action), rather than try to
send the current event record. This solved the problem where Ember was
always passing this data as null to the API.
- Refactor `createTimeslots` to save most data (except form data) locally
within the function rather than storing it in the component
- Refactor `createTimeslots` to pass the array of `newTimeslots` rather
than sequentially passing each `timeslotPojo` individually. The looping
is now done at the `admin` route, which then creates each record
individually.
- Refactor `timeslots-form.js` to show either the existing timeslots
with a 'start over' button (to call `deleteTimeslots`) or the form with
a 'build schedule' button (to call `createTimeslots`).